// client/src/core/dailyRules.generated.ts
// Generated by zkube-codegen. Do not edit.
export const DAILY_PAIR_COUNT = 160 as const;
export const DAILY_PAIR_SELECTION_SEED = [122, 107, 117, 98, 101, 45, 100, 97, 105, 108, 121, 45, 112, 111, 111, 108, 45, 118, 48, 49, 45, 112, 117, 98, 108, 105, 99, 45, 115, 101, 101, 100] as const;
export const DAILY_THEMES = [{ kind: 0, value: 0 }, { kind: 1, value: 2 }, { kind: 1, value: 3 }, { kind: 4, value: 2 }, { kind: 4, value: 3 }, { kind: 2, value: 1 }, { kind: 2, value: 2 }, { kind: 2, value: 3 }, { kind: 2, value: 4 }, { kind: 6, value: 0 }, { kind: 7, value: 0 }, { kind: 8, value: 0 }, { kind: 17, value: 7 }, { kind: 17, value: 8 }, { kind: 18, value: 3 }, { kind: 18, value: 4 }] as const;